+++ Runtime/Scripts/Schema/FakeSchema/Mesh.cs
@@ -1,14 +1,12 @@
 // SPDX-FileCopyrightText: 2023 Unity Technologies and the glTFast authors
 // SPDX-License-Identifier: Apache-2.0
 
-#if DRACO_UNITY
-
-namespace GLTFast.FakeSchema {
+namespace GLTFast.FakeSchema
+{
 
     [System.Serializable]
-    class Mesh : NamedObject {
+    class Mesh : NamedObject
+    {
         public MeshPrimitive[] primitives;
     }
 }
-
-#endif